[NEW] libfx2: Add libfx2 module, make SDCC not delete files
At the time of writing SDCC's SVN trunk fails to compile any 8051 code
that uses the libfx2 Make rules, so I'm maintaining a Git branch
mirroring the SDCC SVN that is called "non-deleter" that is pinned to
the 4.2.0 release.  MacPorts seems to be holding back their SDCC package
at a similar version, presumably for the same reason(??)

I will be investigating why SCCC's C preprocessor is deleting all of my
code at some point in the future and will hopefully collaborate with
upstream to ensure this doesn't happen with libfx2 in the future, but
for now I at least have a working toolchain.