[NEW] efi-booter: Check in initial unfinished SPI flash drive code
I usually don't check in non-working code to git, but this is really
getting out of hand...

The efi-booter FX2 firmware is the code that serves up a USB mass
storage device using an external SPI flash chip as a backing store (well
it will once it's finished).  The target computer uses this drive to
load the targetdisk UEFI runtime before reflashing its half of the FX2
dongle to ready it for our device-FIFO-device mass storage sandwich.

What this all means will all be revealed in future commits, I can assure
you.  ;)